--- ml_logging/__init__.py
@@ -1,1 +1,0 @@
-from .mlflow_utils import get_tracking_uri
