+++ squid/server/infra/clean-infra.sh
