+++ squid/server/infra/minio/create-bucket.sh
