+++ squid/server/infra/postgres/init.sql
