+++ squid/server/operations.py
